switch to javascript instead of typescript.
