+++ components/log.d.ts
@@ -3,11 +3,8 @@
 // Licensed under the MIT License. See License file under the project root for license information.
 //-----------------------------------------------------------------------------
 
-import { CompilerOptions } from "typescript";
 import { Transform } from "stream";
 
-declare namespace tsc {
-    function compile(options?: CompilerOptions): NodeJS.WritableStream & NodeJS.ReadableStream;
-}
+declare const log: () => Transform;
 
-export = tsc;
+export = log;
