--- components/tsc/package.json
@@ -1,5 +1,0 @@
-{
-    "name": "tsc",
-    "version": "1.0.0",
-    "main": "tsc.js"
-}
