--- components/tsc/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES2017",
-        "declaration": false,
-        "allowJs": true
-    }
-}
