--- components/wix/package.json
@@ -1,5 +1,0 @@
-{
-    "name": "wix",
-    "version": "1.0.0",
-    "main": "wix.js"
-}
