+++ processors/electron/pack.d.ts
