+++ processors/typescript.d.ts
@@ -3,6 +3,9 @@
 // Licensed under the MIT License. See License file under the project root for license information.
 //-----------------------------------------------------------------------------
 
-export declare function normalizeGlobs(...globs: Array<string>): Array<string>;
-
-export declare function toGlobs(globlike: import("./configs").GlobLike, exts?: string | Array<string>): Array<string>;
+declare function constructProcessor(
+    config,
+    buildTarget: IBuildTaget,
+    buildInfos: IBuildInfos,
+    packageJson: IPackageConfig): NodeJS.ReadableStream & NodeJS.WritableStream;
+export = constructProcessor;
